+++ proxmoxtf/resource_virtual_environment_vm_test.go
@@ -89,6 +89,7 @@ func TestResourceVirtualEnvironmentVMSchema(t *testing.T) {
 
 	testOptionalArguments(t, agentSchema, []string{
 		mkResourceVirtualEnvironmentVMAgentEnabled,
+		mkResourceVirtualEnvironmentVMAgentTimeout,
 		mkResourceVirtualEnvironmentVMAgentTrim,
 		mkResourceVirtualEnvironmentVMAgentType,
 	})